A3 with LCD string and Keypad
PLEASE READ: At this state, the code has everything needed to complete the main assignment: implementation of LCD control and Keypad. The lcd and keypad modules are designed to be easily implemented and allow use of any viable GPIO pins with some exceptions: GPIO pins must be in order, and the lcd commands require the lower 4 (0,1,2,3) of a GPIO port to be used.